+++ examples/react-template/screens/Calendar.tsx
@@ -1,4 +1,4 @@
-import { Button, Calendar, ChangeEventCalendar, Column, Columns, Section, Text } from '@trilogy-ds/react'
+import { Button, Calendar, ChangeEventCalendar, Section, Spacer, Text, View } from '@trilogy-ds/react'
 import { useState } from 'react'
 
 const minDate = new Date(2025, 9, 10)
@@ -11,36 +11,40 @@ export const CalendarScreen = (): JSX.Element => {
 
   return (
     <Section>
-      <Columns align='ALIGNED_CENTER'>
-        <Column narrow>
-          <Text>Simple Calendar</Text>
-          <Button onClick={() => setValue(new Date(2025, 9, 11))} variant='CONVERSION'>
-            Set Date {new Date(2025, 9, 11).toLocaleDateString()}
-          </Button>
-          <Calendar
-            onMonthChange={(e) => console.log(e)}
-            disabledDates={disabledDates}
-            maxDate={maxDate}
-            minDate={minDate}
-            value={value}
-            onChange={(e) => {
-              setValue(e)
-            }}
-          />
-        </Column>
-        <Column narrow>
-          <Text>Range Calendar</Text>
-          <Button onClick={() => setValues([new Date(2025, 9, 11), new Date(2025, 9, 20)])} variant='CONVERSION'>
-            Set Date [{new Date(2025, 9, 11).toLocaleDateString()}, {new Date(2025, 9, 20).toLocaleDateString()} ]
-          </Button>
-          <Calendar
-            value={values}
-            onChange={(e) => {
-              setValues(e)
-            }}
-          />
-        </Column>
-      </Columns>
+      <View>
+        <Text>Simple Calendar</Text>
+        <Button onClick={() => setValue(new Date(2025, 9, 11))} variant='CONVERSION'>
+          {`Set Date ${new Date(2025, 9, 11).toLocaleDateString()}`}
+        </Button>
+
+        <Spacer size={16} />
+
+        <Calendar
+          onMonthChange={(e) => console.log(e)}
+          disabledDates={disabledDates}
+          maxDate={maxDate}
+          minDate={minDate}
+          value={value}
+          onChange={(e) => {
+            setValue(e)
+          }}
+        />
+      </View>
+      <View>
+        <Text>Range Calendar</Text>
+        <Button onClick={() => setValues([new Date(2025, 9, 11), new Date(2025, 9, 20)])} variant='CONVERSION'>
+          {`Set Date [${new Date(2025, 9, 11).toLocaleDateString()}, ${new Date(2025, 9, 20).toLocaleDateString()} ]`}
+        </Button>
+
+        <Spacer size={16} />
+
+        <Calendar
+          value={values}
+          onChange={(e) => {
+            setValues(e)
+          }}
+        />
+      </View>
     </Section>
   )
 }