+++ examples/react-template/screens/Modal.tsx
@@ -20,12 +20,10 @@ import {
   Text,
   Title,
   TitleLevels,
-  useTrilogyContext,
 } from '@trilogy-ds/react'
 import { useState } from 'react'
 
 export const ModalScreen = (): JSX.Element => {
-  const trilogy = useTrilogyContext()
   const [openModal1, setOpenModal1] = useState(false)
   const [openModal2, setOpenModal2] = useState(false)
   const [openModal3, setOpenModal3] = useState(false)
@@ -97,9 +95,12 @@ export const ModalScreen = (): JSX.Element => {
         <Modal
           title='Hello'
           trigger={
-            <Button onClick={() => setOpenModal2(true)}>
-              <Icon name={IconName.EYE} size={IconSize.LARGE} color={IconColor.MAIN} />
-            </Button>
+            <Icon
+              name={IconName.EYE}
+              size={IconSize.LARGE}
+              color={IconColor.MAIN}
+              onClick={() => setOpenModal2(true)}
+            />
           }
           active={openModal2}
           onClose={() => setOpenModal2(false)}
@@ -118,9 +119,12 @@ export const ModalScreen = (): JSX.Element => {
         <Modal
           title='Hello'
           trigger={
-            <Button onClick={() => setOpenModal3(true)}>
-              <Icon name={IconName.EYE} size={IconSize.LARGE} color={IconColor.MAIN} />
-            </Button>
+            <Icon
+              name={IconName.EYE}
+              size={IconSize.LARGE}
+              color={IconColor.MAIN}
+              onClick={() => setOpenModal3(true)}
+            />
           }
           active={openModal3}
           onClose={() => setOpenModal3(false)}