+++ packages/react/components/icon/Icon.native.tsx
@@ -7,8 +7,8 @@ import { isAndroid, isIOS } from '@/helpers/device/index.native'
 import { getAlignStyle } from '@/objects/facets/Alignable'
 import { getColorStyle, TrilogyColor, TrilogyColorValues } from '@/objects/facets/Color/index.native'
 import React, { useContext } from 'react'
-import ContentLoader, { Circle } from 'react-content-loader/native'
 import { StyleSheet, TouchableOpacity, View } from 'react-native'
+import { Skeleton } from '../skeleton'
 import { WithLocalSvg } from 'react-native-svg/css'
 
 /**
@@ -117,17 +117,13 @@ const Icon = React.forwardRef<IconNativeRef, IconProps>(
     })
 
     const IconSkeleton = (): JSX.Element => (
-      <ContentLoader style={styles.skeleton} {...others}>
-        <View style={{ opacity: 0 }} />
-        {isAndroid && (
-          <View>
-            {(size === IconSize.HUGE && <Circle cx='50' cy='50' r='50' />) ||
-              (size === IconSize.LARGE && <Circle cx='33' cy='33' r='33' />) ||
-              (size === IconSize.MEDIUM && <Circle cx='23' cy='23' r='23' />) ||
-              (size === IconSize.SMALL && <Circle cx='15' cy='15' r='15' />) || <Circle cx='15' cy='15' r='15' />}
-          </View>
-        )}
-      </ContentLoader>
+      <Skeleton
+        style={styles.skeleton}
+        width={circledWidth}
+        height={circledWidth}
+        borderRadius={iconSkeletonRadius}
+        testID="skeleton-id"
+      />
     )
 
     if (skeleton) {