+++ packages/react/components/skeleton/index.ts
@@ -1,0 +1,5 @@
+import Skeleton from './Skeleton.native'
+
+export * from './SkeletonProps'
+
+export { Skeleton }